[refactor/#10] - IconButton에 React.FC<SVGProps>를 넘겨주도록 변경
