+++ src/constants/urls.ts
@@ -1,5 +1,5 @@
 export const URLS = {
   HOME: "/",
-  LOGIN: "/login", // 임시
+  LOGIN: "/", // 임시
   HOMEPAGE: "https://www.hansung.ac.kr/hansung/index.do",
 } as const;
